# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/create": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreatedUserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建用户（管理员以上）",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/list": {
      "post": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserListQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserListDataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "用户列表（管理员以上）",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/profile/detail": {
      "post": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserSafeDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取个人详情",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/profile/update": {
      "post": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserSafeDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改个人资料",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/profile/change-password": {
      "post": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改密码",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/detail": {
      "post": {
        "operationId": "UserController_findOne",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserSafeDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取指定用户详情（管理员以上）",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/update": {
      "post": {
        "operationId": "UserController_adminUpdateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserSafeDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新用户信息（管理员以上，需高于目标用户角色）",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/update-status": {
      "post": {
        "operationId": "UserController_updateStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改用户状态（管理员以上，需高于目标用户角色）",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/reset-password": {
      "post": {
        "operationId": "UserController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResetPasswordDataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "重置用户密码（管理员以上，需高于目标用户角色）",
        "tags": [
          "User - 用户"
        ]
      }
    },
    "/user/delete": {
      "post": {
        "operationId": "UserController_remove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除用户（软删除，需高于目标用户角色）",
        "tags": [
          "User - 用户"
        ]
      }
    },
...
  },
  "info": {
    "title": "Quant Server API",
    "description": "量化交易后端接口文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "ResponseModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "default": 0
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "default": ""
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
...
      "CreatedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "account": {
            "type": "string"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN",
              "USER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DEACTIVATED",
              "DELETED"
            ]
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "wechat": {
            "type": "string",
            "nullable": true
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "backtestQuota": {
            "type": "number"
          },
          "watchlistLimit": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "initialPassword": {
            "type": "string",
            "description": "新建用户初始密码（仅本次返回）"
          }
        },
        "required": [
          "id",
          "account",
          "role",
          "status",
          "backtestQuota",
          "watchlistLimit",
          "createdAt",
          "updatedAt",
          "initialPassword"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "example": "zhangsan",
            "description": "登录账号"
          },
          "nickname": {
            "type": "string",
            "example": "张三",
            "description": "昵称"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN",
              "USER"
            ],
            "description": "用户角色（默认 USER）"
          }
        },
        "required": [
          "account",
          "nickname"
        ]
      },
      "UserSafeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "account": {
            "type": "string"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN",
              "USER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DEACTIVATED",
              "DELETED"
            ]
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "wechat": {
            "type": "string",
            "nullable": true
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "backtestQuota": {
            "type": "number"
          },
          "watchlistLimit": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account",
          "role",
          "status",
          "backtestQuota",
          "watchlistLimit",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserListDataDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSafeDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "items"
        ]
      },
      "UserListQueryDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码，从1开始",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数",
            "default": 20
          },
          "account": {
            "type": "string",
            "description": "账号模糊搜索"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DEACTIVATED",
              "DELETED"
            ],
            "description": "用户状态筛选"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN",
              "USER"
            ],
            "description": "用户角色筛选"
          }
        }
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "李四",
            "description": "昵称"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "邮箱"
          },
          "wechat": {
            "type": "string",
            "example": "wx_12345",
            "description": "微信号"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "旧密码"
          },
          "newPassword": {
            "type": "string",
            "description": "新密码（至少8位）"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "UserIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "用户 ID"
          }
        },
        "required": [
          "id"
        ]
      },
      "AdminUpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "用户 ID"
          },
          "nickname": {
            "type": "string",
            "example": "张三",
            "description": "昵称"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "邮箱"
          },
          "wechat": {
            "type": "string",
            "example": "wx_12345",
            "description": "微信号"
          },
          "backtestQuota": {
            "type": "number",
            "example": 5,
            "description": "回测任务数量限制"
          },
          "watchlistLimit": {
            "type": "number",
            "example": 20,
            "description": "监控股票数量限制"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "用户 ID"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DEACTIVATED"
            ],
            "description": "用户状态"
          }
        },
        "required": [
          "id",
          "status"
        ]
      },
      "ResetPasswordDataDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
    }
  }
}